Renamed the components files to match conventions
