+++ client/components/AuthForm.js
@@ -1,6 +1,6 @@
-import React from 'react'
-import {connect} from 'react-redux'
-import PropTypes from 'prop-types'
+import React from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react'
+import {connect} from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react-redux'
+import PropTypes from '../../../../Library/Caches/typescript/2.9/node_modules/@types/prop-types'
 import {auth} from '../store'
 
 /**
