+++ client/components/Home.js
@@ -1,5 +1,5 @@
-import React from 'react'
-import Navbar from './Navbar'
+import React from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react'
+import Navbar from './navbar'
 import ProductList from './ProductList'
 
 const Home = () =>{
