+++ client/components/ProductList.js
@@ -1,5 +1,5 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux';
+import React, { Component } from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react'
+import { connect } from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react-redux';
 import ProductCard from './ProductCard'
 import { fetchAllProducts } from '../store/product'
