+++ client/components/UserHome.js
@@ -1,6 +1,6 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import {connect} from 'react-redux'
+import React from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react'
+import PropTypes from '../../../../Library/Caches/typescript/2.9/node_modules/@types/prop-types'
+import {connect} from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react-redux'
 
 /**
  * COMPONENT
