+++ client/components/index.js
@@ -4,5 +4,5 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
+export {default as UserHome} from './UserHome'
+export {Login, Signup} from './AuthForm'
