+++ client/routes.js
@@ -1,7 +1,7 @@
-import React, {Component} from 'react'
-import {connect} from 'react-redux'
-import {withRouter, Route, Switch} from 'react-router-dom'
-import PropTypes from 'prop-types'
+import React, {Component} from '../../../Library/Caches/typescript/2.9/node_modules/@types/react'
+import {connect} from '../../../Library/Caches/typescript/2.9/node_modules/@types/react-redux'
+import {withRouter, Route, Switch} from '../../../Library/Caches/typescript/2.9/node_modules/@types/react-router-dom'
+import PropTypes from '../../../Library/Caches/typescript/2.9/node_modules/@types/prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
 import Home from './components/Home'
